+++ src/main/java/steamcraft/common/tiles/TileSteamBoiler.java
@@ -1,5 +1,5 @@
 /**
- * This class was created by BrassGoggledCoders modding team. 
+ * This class was created by BrassGoggledCoders modding team.
  * This class is available as part of the Steamcraft 2 Mod for Minecraft.
  *
  * Steamcraft 2 is open-source and is distributed under the MMPL v1.0 License.
@@ -8,7 +8,7 @@
  * Steamcraft 2 is based on the original Steamcraft Mod created by Proloe.
  * Steamcraft (c) Proloe 2011
  * (http://www.minecraftforum.net/topic/251532-181-steamcraft-source-code-releasedmlv054wip/)
- * 
+ *
  */
 package steamcraft.common.tiles;
 
@@ -29,7 +29,7 @@ import boilerplate.common.baseclasses.BaseTileWithInventory;
 
 /**
  * @author Decebaldecebal
- * 
+ *
  */
 public class TileSteamBoiler extends BaseTileWithInventory implements IFluidHandler
 {
@@ -45,7 +45,7 @@ public class TileSteamBoiler extends BaseTileWithInventory implements IFluidHand
 
 	public TileSteamBoiler()
 	{
-		super((byte) 3);
+		super(3);
 
 		this.waterTank = new FluidTank(new FluidStack(FluidRegistry.WATER, 0), 5000);
 		this.steamTank = new FluidTank(new FluidStack(FluidRegistry.getFluid("steam"), 0), 10000);
